web-animations/interfaces/KeyframeEffect/mutable-effects.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Tests for modifying animation effects</title>
<link rel="help" href="https://w3c.github.io/web-animations/#the-keyframeeffect-interfaces">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<body>
<script>
'use strict';

test(function(t) {
  document.timeline.playbackRate = 0;
  document.timeline.currentTime = 0;
  var element = createDiv(t);
  var effect = new KeyframeEffect(element, null, 1000);
  var animation = new Animation(effect);
  animation.startTime = -1000;
  assert_equals(animation.currentTime, 1000);
  animation.effect.timing.duration = 1500;
  document.timeline.currentTime = 2000;
  assert_equals(animation.startTime, -1000);
  assert_equals(animation.currentTime, 1500);
}, 'Extending effect should not cause a change in start time.');

</script>
</body>
